added module for instructor add and update

+++ backend/LecturerCourseDemo/src/main/java/com/crestasom/lecturercoursedemo/repo/InstructorRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.crestasom.lecturercoursedemo.model.Instructor;
 
-public interface InstructorRepo extends JpaRepository<Instructor, Long> {
+public interface InstructorRepo extends JpaRepository<Instructor, Integer> {
 	Instructor findByUserName(String userName);
 }
